+++ CHANGELOG.md
@@ -7,6 +7,7 @@ IMPROVEMENTS:
 - Resource compute_instance: clearify the default state value #538
 - Change zone for sks acceptance tests
 - Migrate Private Network resource/datasource to framework #542
+- feat: add uri to every dbass service #567
 
 BUG FIXES:
 - fix: typo in security group rule import doc #559